requirements.txt
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
